Fixed all bugs, finished this test
